requirements.txt
python-telegram-bot==21.*
python-dotenv>=1.0.0   # ixtiyoriy, .env fayl uchun